ci(ccusage): use equal perf fixture defaults
Raise the Codex large fixture default from 256 MiB to 1024 MiB so CI compares Claude and Codex against the same generated input size unless explicitly overridden.

Move the performance script assertions into in-source Vitest blocks beside the script logic. The ccusage Vitest config now includes only these performance scripts as script in-source tests, avoiding unrelated Bun-only scripts that are not currently Vite-resolvable.

Validated with pnpm --filter ccusage test scripts/compare-pr-performance.ts scripts/generate-large-fixture.ts, pnpm run format, pnpm typecheck, pnpm --filter ccusage run build, and pnpm run test.
