+++ apps/ccusage/vitest.config.ts
@@ -4,7 +4,11 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
 	test: {
 		watch: false,
-		includeSource: ['src/**/*.{js,ts}'],
+		includeSource: [
+			'src/**/*.{js,ts}',
+			'scripts/compare-pr-performance.ts',
+			'scripts/generate-large-fixture.ts',
+		],
 		globals: true,
 	},
 	plugins: [
